+++ plugins/commentable-html/dev/skill/tools/validate/cmhval/contrast.py
@@ -483,7 +483,7 @@ class _StyleScanner(_browser_boundaries.BrowserBoundaries):
 
     def handle_endtag(self, tag):
         tag = self._browser_tag(tag)
-        for i in range(len(self._els) - 1, -1, -1):
+        for i in range(len(self._els) - 1, self._end_tag_floor(tag) - 1, -1):
             if self._els[i] == tag:
                 self._truncate_stacks(i)
                 return